+++ packages/plugin-packages/mashroom-portal-remote-app-registry-k8s/src/js/jobs/ScanK8SPortalRemoteAppsBackgroundJob.ts
@@ -1,5 +1,3 @@
-
-import {KubeConfig, CoreV1Api} from '@kubernetes/client-node';
 import url from 'url';
 import request from 'request';
 import context from '../context';
@@ -10,24 +8,22 @@ import {
     MashroomPluginPackageDefinition
 } from "@mashroom/mashroom/type-definitions";
 import {MashroomPortalApp, MashroomPortalProxyDefinitions} from "@mashroom/mashroom-portal/type-definitions";
-import {KubernetesService, ScanBackgroundJob} from "../../../type-definitions";
+import {KubernetesConnector, KubernetesService, ScanBackgroundJob} from "../../../type-definitions";
 
 export default class ScanK8SPortalRemoteAppsBackgroundJob implements ScanBackgroundJob {
 
-    private k8sApi?: CoreV1Api;
     private timeout?: NodeJS.Timeout;
     private readonly serviceNameFilter: RegExp;
     private readonly logger: MashroomLogger;
 
-    constructor(private k8sNamespaces: Array<string>, serviceNameFilterStr: string, private scanPeriodSec: number, private refreshIntervalSec: number, private accessViaClusterIP: boolean, loggerFactory: MashroomLoggerFactory) {
-        this.serviceNameFilter = new RegExp(serviceNameFilterStr);
+    constructor(private k8sNamespaces: Array<string>, serviceNameFilterStr: string, private scanPeriodSec: number, private refreshIntervalSec: number,
+                private accessViaClusterIP: boolean, private kubernetesConnector: KubernetesConnector, loggerFactory: MashroomLoggerFactory) {
+        this.serviceNameFilter = new RegExp(serviceNameFilterStr, 'i');
         this.logger = loggerFactory('mashroom.portal.remoteAppRegistryK8s');
     }
 
     start(): void {
-        if (!this.k8sApi) {
-            this.setupK8sApi();
-        }
+        this.kubernetesConnector.init();
         this.scanKubernetesServices();
         this.timeout = setInterval(() => this.scanKubernetesServices(), this.scanPeriodSec * 1000);
     }
@@ -38,18 +34,7 @@ export default class ScanK8SPortalRemoteAppsBackgroundJob implements ScanBackgro
         }
     }
 
-    private setupK8sApi(): void {
-        const k8sClient = new KubeConfig();
-        // This only works if the Portal runs within a Kubernetes Pod with a valid service account attached
-        k8sClient.loadFromCluster();
-        this.k8sApi = k8sClient.makeApiClient(CoreV1Api);
-    }
-
     private async scanKubernetesServices(): Promise<void> {
-        if (!this.k8sApi) {
-            return;
-        }
-
         this.logger.info('Starting scan of k8s namespaces: ', this.k8sNamespaces);
         context.lastScan = Date.now();
         context.error = null;
@@ -57,29 +42,32 @@ export default class ScanK8SPortalRemoteAppsBackgroundJob implements ScanBackgro
         for (let i = 0; i < this.k8sNamespaces.length; i++) {
             const namespace = this.k8sNamespaces[i];
             try {
-                const res = await this.k8sApi.listNamespacedService(namespace);
-                const serviceItems = res.body.items;
+                const res = await this.kubernetesConnector.listNamespaceServices(namespace);
+                const serviceItems = res.items;
 
                 for (let j = 0; j < serviceItems.length; j++) {
                     const serviceItem = serviceItems[j];
 
                     const name = serviceItem?.metadata?.name;
-                    const ip = serviceItem?.spec?.clusterIP;
-                    const ports = serviceItem?.spec?.ports;
-                    const port = ports && ports.length > 0 ? ports[0].port : undefined;
-
-                    if (name && ip && ip.toLowerCase() !== 'none' && port && port && name.match(this.serviceNameFilter)) {
-                        let service: KubernetesService | null = null;
+                    if (name && name.match(this.serviceNameFilter)) {
                         const existingService = context.registry.getService(name);
+                        const ip = serviceItem?.spec?.clusterIP;
+                        const ports = serviceItem?.spec?.ports;
+                        const port = ports && ports.length > 0 ? ports[0].port : undefined;
+                        const headlessService = !port || !ip || ip.toLowerCase() === 'none';
+                        let service: KubernetesService | undefined;
+
                         if (existingService) {
-                            if (existingService.error || existingService.lastCheck < Date.now() - this.refreshIntervalSec * 1000) {
+                            if (existingService.status === 'Error' || existingService.lastCheck < Date.now() - this.refreshIntervalSec * 1000) {
                                 service = Object.assign({}, existingService, {
+                                    status: 'Checking',
                                     lastCheck: Date.now(),
                                     error: null,
                                 })
                             }
                         } else {
                             const url = this.accessViaClusterIP ? `http://${ip}:${port}` : `http://${name}.${namespace}:${port}`;
+                            this.logger.info(`Adding new Kubernetes service: ${name} (${url})`);
 
                             service = {
                                 name,
@@ -88,24 +76,35 @@ export default class ScanK8SPortalRemoteAppsBackgroundJob implements ScanBackgro
                                 port,
                                 url,
                                 firstSeen: Date.now(),
+                                status: 'Checking',
                                 lastCheck: Date.now(),
                                 error: null,
-                                descriptorFound: false,
                                 foundPortalApps: []
                             };
                         }
 
                         if (service) {
-                            await this.checkServiceForRemotePortalApps(service);
-                            context.registry.addOrUpdateService(service);
+                            if (this.accessViaClusterIP && headlessService) {
+                                service = Object.assign({}, service, {
+                                    status: 'Headless Service',
+                                    foundPortalApps: []
+                                });
+                                context.registry.addOrUpdateService(service);
+                            } else {
+                                context.registry.addOrUpdateService(service);
+                                await this.checkServiceForRemotePortalApps(service);
+                                context.registry.addOrUpdateService(service);
+                            }
                         }
                     }
                 }
             } catch (error) {
-                this.logger.error(`Error during scan of k8s namespace: ${namespace}`, error)
-                context.error = error;
+                this.logger.error(`Error during scan of k8s namespace: ${namespace}`, error);
+                context.error = error.message;
             }
         }
+
+        this.removeAppsNotSeenForALongTime();
     }
 
     private async checkServiceForRemotePortalApps(service: KubernetesService): Promise<void> {
@@ -115,36 +114,62 @@ export default class ScanK8SPortalRemoteAppsBackgroundJob implements ScanBackgro
         let portalApps;
 
         try {
-            packageJson = await this.loadPackageJson(service.url);
-            service.descriptorFound = true;
+            const {found, json} = await this.loadPackageJson(service.url);
+            if (found) {
+                packageJson = json;
+            } else {
+                service.foundPortalApps = [];
+                service.status = 'No Descriptor';
+                return;
+            }
         } catch (error) {
-            service.descriptorFound = false;
-            this.logger.debug(`Fetching remote portal app info for Kubernetes service ${service.name} failed!`);
+            service.status = 'Error';
+            service.error = error.message;
+            service.foundPortalApps = [];
+            this.logger.error(`Error checking Kubernetes service ${service.name} failed!`, error);
             return;
         }
 
         try {
             portalApps = this.processPackageJson(packageJson, service.url, service.name);
+            this.logger.info(`Registered portal apps for Kubernetes service: ${service.name}:`, portalApps);
         } catch (error) {
-            this.logger.error(`Processing remote portal app info for Kubernetes service ${service.name} failed!`, error);
+            service.status = 'Error';
             service.error = error.message;
+            service.foundPortalApps = [];
+            this.logger.error(`Processing remote portal app info for Kubernetes service ${service.name} failed!`, error);
             return;
         }
 
+        service.status = 'Valid';
         service.foundPortalApps = portalApps;
     }
 
-    private async loadPackageJson(serviceUrl: string): Promise<any> {
+    private async loadPackageJson(serviceUrl: string): Promise<{ found: boolean; json?: any }> {
+        const requestOptions = {
+            url: `${serviceUrl}/package.json`,
+            followRedirect: false,
+            timeout: 3000,
+        };
+
         return new Promise((resolve, reject) => {
-            request.get({ url: `${serviceUrl}/package.json` }, (error, response, body) => {
+            request.get(requestOptions, (error, response, body) => {
                 if (error) {
-                    reject(new Error('No package.json found!'));
+                    reject(error);
+                    return;
+                }
+                if (response.statusCode === 404) {
+                    resolve({found: false});
+                    return;
+                }
+                if (response.statusCode !== 200) {
+                    reject(new Error(`Received HTTP Code: ${response.statusCode}`));
                     return;
                 }
 
                 try {
-                    const packageJson = JSON.parse(body);
-                    resolve(packageJson);
+                    const json = JSON.parse(body);
+                    resolve({found: true, json});
                 } catch (parseError) {
                     reject(new Error('Parsing /package.json failed!\n' + parseError.message));
                 }
@@ -156,14 +181,14 @@ export default class ScanK8SPortalRemoteAppsBackgroundJob implements ScanBackgro
         this.logger.debug(`Processing package.json of Kubernetes service: ${serviceName}`, packageJson);
 
         if (!packageJson || !packageJson.mashroom) {
-            throw new Error(`No mashroom property found in package.json of Kubernetes service: ${serviceName}`);
+            throw new Error('No mashroom property found in package.json');
         }
 
         const mashroomDef: MashroomPluginPackageDefinition = packageJson.mashroom;
         const portalAppDefinitions = mashroomDef.plugins.filter((plugin) => plugin.type === 'portal-app');
 
         if (portalAppDefinitions.length === 0) {
-            throw new Error(`No plugin of type portal-app found in remote portal app of Kubernetes service: ${serviceName}`);
+            throw new Error('No plugin of type portal-app found in remote portal app');
         }
 
         return portalAppDefinitions.map((definition) => this.mapPluginDefinition(packageJson, definition, serviceUrl, serviceName));
@@ -173,17 +198,17 @@ export default class ScanK8SPortalRemoteAppsBackgroundJob implements ScanBackgro
 
         const name = definition.name;
         if (!name) {
-            throw new Error(`Invalid portal app definition: No 'name' attribute! Kubernetes service: ${serviceName}`);
+            throw new Error('Invalid portal app definition: No "name" attribute!');
         }
 
         const globalLaunchFunction = definition.bootstrap;
         if (!globalLaunchFunction) {
-            throw new Error(`Invalid configuration of plugin ${name}: No bootstrap function defined. Kubernetes service: ${serviceName}`);
+            throw new Error('Invalid configuration of plugin ${name}: No bootstrap function defined.');
         }
 
         const resourcesDef = definition.resources;
         if (!resourcesDef) {
-            throw new Error(`Invalid configuration of plugin ${name}: No resources defined. Kubernetes service: ${serviceName}`);
+            throw new Error('Invalid configuration of plugin ${name}: No resources defined.');
         }
 
         const resources = {
@@ -191,7 +216,7 @@ export default class ScanK8SPortalRemoteAppsBackgroundJob implements ScanBackgro
             css: resourcesDef.css,
         };
         if (!resources.js) {
-            throw new Error(`Invalid configuration of plugin ${name}: No resources.js defined. Kubernetes service: ${serviceName}`);
+            throw new Error('Invalid configuration of plugin ${name}: No resources.js defined.');
         }
 
         const sharedResourcesDef = definition.sharedResources;
@@ -255,4 +280,15 @@ export default class ScanK8SPortalRemoteAppsBackgroundJob implements ScanBackgro
 
         return portalApp;
     }
+
+    private removeAppsNotSeenForALongTime(): void {
+        context.registry.services.forEach((service) => {
+            if (service.lastCheck < Date.now() - this.refreshIntervalSec * 1000 * 2.5) {
+                // Not seen for 2.5 check intervals
+                this.logger.info(`Removing portal apps because Kubernetes service ${service.name} is no longer available:`, service.foundPortalApps);
+                context.registry.removeService(service.name);
+            }
+        });
+    }
+
 }