+++ packages/plugin-packages/mashroom-portal-remote-app-registry-k8s/src/js/registry/mashroom-bootstrap-remote-portal-app-registry.ts
@@ -1,6 +1,7 @@
-
 import context from '../context';
 import ScanK8SPortalRemoteAppsBackgroundJob from '../jobs/ScanK8SPortalRemoteAppsBackgroundJob';
+import KubernetesConnector from '../k8s/KubernetesConnector';
+import DummyKubernetesConnector from '../k8s/DummyKubernetesConnector';
 
 import {MashroomRemotePortalAppRegistryBootstrapFunction} from '@mashroom/mashroom-portal/type-definitions';
 
@@ -8,12 +9,16 @@ const bootstrap: MashroomRemotePortalAppRegistryBootstrapFunction = async (plugi
     const pluginContext = contextHolder.getPluginContext();
     const {k8sNamespaces, scanPeriodSec, refreshIntervalSec, serviceNameFilter, accessViaClusterIP} = pluginConfig;
 
+    const kubernetesConnector = process.env.DUMMY_K8S_CONNECTOR ? new DummyKubernetesConnector() : new KubernetesConnector();
 
-    const backgroundJob = new ScanK8SPortalRemoteAppsBackgroundJob(k8sNamespaces, serviceNameFilter, scanPeriodSec, refreshIntervalSec, accessViaClusterIP, pluginContext.loggerFactory);
+    const backgroundJob = new ScanK8SPortalRemoteAppsBackgroundJob(k8sNamespaces, serviceNameFilter, scanPeriodSec, refreshIntervalSec,
+        accessViaClusterIP, kubernetesConnector, pluginContext.loggerFactory);
     backgroundJob.start();
     pluginContext.services.core.pluginService.onUnloadOnce(pluginName, () => {
         backgroundJob.stop();
     });
+
+    context.serviceNameFilter = serviceNameFilter;
 
     return context.registry;
 };