+++ packages/plugin-packages/mashroom-portal-remote-app-registry-k8s/src/js/webapp/routes/services.ts
@@ -1,30 +1,52 @@
 import context from '../../context';
 
 import {Request, Response} from "express";
-import {ServicesRenderModel} from "../../../../type-definitions";
+import {KubernetesServiceStatus, ServicesRenderModel} from "../../../../type-definitions";
 
 const formatDate = (ts: number): string => {
     return new Date(ts).toISOString().replace(/T/, ' ').replace(/\..+/, '')
+};
+
+const getRowClass = (status: KubernetesServiceStatus): string => {
+    switch (status) {
+        case 'Error':
+            return 'row-error';
+        case 'Headless Service':
+        case 'No Descriptor':
+            return 'row-ignored';
+        default:
+            return '';
+    }
+};
+
+const getStatusClass = (status: KubernetesServiceStatus): string => {
+    switch (status) {
+        case 'Error':
+            return 'error';
+        case 'Checking':
+            return 'checking';
+        default:
+            return '';
+    }
 };
 
 export default (request: Request, response: Response) => {
 
     const model: ServicesRenderModel = {
         baseUrl: request.baseUrl,
-        error: context.error,
+        scanError: context.error,
         lastScan: formatDate(context.lastScan),
-        totalServices: String(context.registry.services.length),
+        serviceNameFilter: context.serviceNameFilter,
         services: context.registry.services
-            .filter((service) => service.descriptorFound)
             .map((service) => ({
                 name: service.name,
                 namespace: service.namespace,
                 url: service.url,
-                status: service.error ? 'Error' : 'Registered',
+                status: service.status !== 'Error' ? service.status : `Error: ${service.error}`,
                 portalApps: service.foundPortalApps.map((app) => `${app.name} (${app.version})`).join(', '),
                 lastCheck: formatDate(service.lastCheck),
-                rowClass: service.error ? 'row-error' : '',
-                statusClass: service.error ? 'error' : 'registered',
+                rowClass: getRowClass(service.status),
+                statusClass: getStatusClass(service.status),
             }))
     };