+++ packages/plugin-packages/mashroom-portal-remote-app-registry-k8s/src/js/webapp/webapp.ts
@@ -1,4 +1,3 @@
-
 import path from 'path';
 import express from 'express';
 import exphbs from 'express-handlebars';
@@ -6,8 +5,7 @@ import services from './routes/services';
 
 const app = express();
 
-app.engine('handlebars', exphbs({
-}));
+app.engine('handlebars', exphbs({}));
 app.set('view engine', 'handlebars');
 app.set('views', path.resolve(__dirname, '../../views'));